+++ CHANGELOG.md
@@ -86,8 +86,8 @@ The new weather input acquires current and future weather conditions from openwe
  - Refactor Edge detection Input
  - Refactor method implementation from single large method into multiple small classes
  - Changed duration method start- and end-time handling
- - Port Math controllers to Functions: Equation, Average (Single/Multi), Sum (Single/Multi)
- - Deprecate Math controllers: Equation, Average (Single), Average (Single/Multi), Sum (Single/Multi)
+ - Port Math controllers to Functions: Equation (Single/Multi), Difference, Statistics (Multi), Average (Single/Multi), Sum (Single/Multi), Wet-Bulb Humidity
+ - Deprecate Math controllers
 
 
 ## 8.8.8 (2020-10-30)